feat(Audit): begin of feats to CRM. adds automatic calculation of ClientAuditMetrics through events

+++ src/main/java/com/rafael/nailspro/webapp/infrastructure/dto/admin/client/ClientAppointmentDTO.java
@@ -3,6 +3,7 @@ package com.rafael.nailspro.webapp.infrastructure.dto.admin.client;
 import com.rafael.nailspro.webapp.domain.enums.AppointmentStatus;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 import java.util.List;
 
@@ -18,7 +19,7 @@ public record ClientAppointmentDTO(
 
         String mainServiceName,
         List<String> addOnServiceNames,
-        Integer totalValue,
+        BigDecimal totalValue,
 
         String observations
 ) {
