+++ src/main/java/com/rafael/nailspro/webapp/infrastructure/dto/appointment/ProfessionalAppointmentScheduleDTO.java
@@ -3,6 +3,7 @@ package com.rafael.nailspro.webapp.infrastructure.dto.appointment;
 import com.rafael.nailspro.webapp.domain.enums.AppointmentStatus;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 @Builder
@@ -15,7 +16,7 @@ public record ProfessionalAppointmentScheduleDTO(Long appointmentId,
                                                  Integer clientCanceledAppointments,
 
                                                  AppointmentStatus status,
-                                                 Integer totalValue,
+                                                 BigDecimal totalValue,
                                                  String observations,
 
                                                  ZonedDateTime startDate,
